# docs/openapi.yaml
openapi: 3.0.3
info:
  title: RivicQ — Encryption as a Service API
  description: |
    ## Quantum-Safe Cryptographic Bill of Materials Management
    
    CryptoBOM SaaS provides comprehensive cryptographic asset discovery, analysis, 
    and compliance management with quantum computing integration.
    
    ### Features
    - Multi-provider asset discovery
    - Quantum vulnerability assessment
    - Compliance framework support (NIST, ISO, BSI, SOX, PCI-DSS)
    - DevSecOps pipeline integration
    - Post-quantum migration planning
    
    ### Editions
    - **OSS Edition**: Basic CBOM with mock quantum providers
    - **Enterprise Edition**: Full IBM Quantum, KIPU Q-CTRL, multi-language SDKs
    
    ### IP Notice
    All intellectual property rights are the exclusive property of **RivicQ GmbH** 
    under German and international law.
  version: 1.3.0
  contact:
    name: RivicQ GmbH
    url: https://rivicq.com
    email: enterprise@rivicq.com
  license:
    name: RivicQ Open Source License (R-OSL)
    url: https://github.com/rivic-q/cryptobom-saas/blob/main/LICENSE
  termsOfService: https://rivicq.com/terms
servers:
  - url: http://localhost:9090
    description: Local development server
  - url: https://api.rivicq.oss.rivicq.com
    description: OSS Edition cloud endpoint
  - url: https://api.rivicq.enterprise.rivicq.com
    description: Enterprise Edition cloud endpoint
...
paths:
...
  /api/v1/engine/migration-plan:
    post:
      tags:
        - Quantum
      summary: Generate post-quantum migration plan
      operationId: migrationPlan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MigrationPlanRequest'
      responses:
        '200':
          description: Migration plan generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MigrationPlan'
components:
  schemas:
...
    MigrationPlanRequest:
      type: object
      properties:
        assets:
          type: array
          items:
            type: string
        targetFramework:
          type: string
    MigrationPlan:
      type: object
      properties:
        phases:
          type: array
          items:
            type: object
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
  - OAuth2:
      - read
      - write
...