# internal/api/enterprise/quantum_attestation.go
package enterprise

import (
	"context"
	"encoding/json"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
	"github.com/rivic-q/cryptobom-saas/internal/database"
	"github.com/rivic-q/cryptobom-saas/internal/quantum/provider"
	"github.com/sirupsen/logrus"
)

type QuantumAttestationHandler struct {
	db        *database.EnterpriseDB
	logger    *logrus.Logger
	providers *provider.Registry
}

func NewQuantumAttestationHandler(db *database.EnterpriseDB, logger *logrus.Logger, registry *provider.Registry) *QuantumAttestationHandler {
	if registry == nil {
		registry = provider.NewRegistry()
	}
	return &QuantumAttestationHandler{
		db:        db,
		logger:    logger,
		providers: registry,
	}
}

func (h *QuantumAttestationHandler) SetupRoutes(router *gin.RouterGroup) {
	quantum := router.Group("/quantum")
	{
		quantum.GET("/attestations", h.ListAttestations)
		quantum.POST("/attestations", h.CreateAttestation)
		quantum.GET("/attestations/:id", h.GetAttestation)
		quantum.POST("/attestations/:id/verify", h.VerifyAttestation)
		quantum.POST("/attestations/:id/refresh", h.RefreshAttestation)

		quantum.GET("/networks", h.ListQuantumNetworks)
		quantum.GET("/providers", h.ListQuantumProviders)
		quantum.GET("/providers/:name/status", h.GetQuantumProviderStatus)

		quantum.GET("/readiness", h.GetQuantumReadiness)
		quantum.GET("/migration", h.GetMigrationPlan)

		quantum.GET("/algorithms", h.ListPQCAlgorithms)
		quantum.POST("/algorithms/migrate", h.MigrateAlgorithm)
	}
}

type QuantumAttestation struct {
	ID              uuid.UUID              `json:"id"`
	TenantID        uuid.UUID              `json:"tenant_id"`
	AssetID         *uuid.UUID             `json:"asset_id"`
	AttestationType string                 `json:"attestation_type"`
	QuantumNetwork  string                 `json:"quantum_network"`
	QuantumProvider string                 `json:"quantum_provider"`
	QubitCount      int                    `json:"qubit_count"`
	ErrorRate       float64                `json:"error_rate"`
	Status          string                 `json:"status"`
	AttestationData map[string]interface{} `json:"attestation_data"`
	AttestedAt      *time.Time             `json:"attested_at"`
	ExpiresAt       *time.Time             `json:"expires_at"`
}

type QuantumNetwork struct {
	ID        string  `json:"id"`
	Name      string  `json:"name"`
	Provider  string  `json:"provider"`
	Status    string  `json:"status"`
	Qubits    int     `json:"qubits"`
	ErrorRate float64 `json:"error_rate"`
	Fidelity  float64 `json:"fidelity"`
	Region    string  `json:"region"`
}

type QuantumReadiness struct {
	OverallScore      int            `json:"overall_score"`
	QuantumSafeAssets int            `json:"quantum_safe_assets"`
	AtRiskAssets      int            `json:"at_risk_assets"`
	MigrationProgress int            `json:"migration_progress"`
	Recommendations   []string       `json:"recommendations"`
	ByCategory        map[string]int `json:"by_category"`
	ByAlgorithm       map[string]int `json:"by_algorithm"`
}

func (h *QuantumAttestationHandler) ListAttestations(c *gin.Context) {
	tenantID := c.GetHeader("X-Tenant-ID")
	status := c.Query("status")

	query := `
		SELECT id, tenant_id, asset_id, attestation_type, quantum_network, quantum_provider,
		       qubit_count, error_rate, status, attestation_data, attested_at, expires_at
		FROM quantum_attestations 
		WHERE tenant_id = $1
	`
	args := []interface{}{tenantID}

	if status != "" {
		query += " AND status = $2"
		args = append(args, status)
	}

	rows, err := h.db.Query(query, args...)
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to list attestations"})
		return
	}
	defer func() { _ = rows.Close() }()

	var attestations []QuantumAttestation
	for rows.Next() {
		var attestation QuantumAttestation
		var attestationData []byte
		err := rows.Scan(
			&attestation.ID, &attestation.TenantID, &attestation.AssetID,
			&attestation.AttestationType, &attestation.QuantumNetwork, &attestation.QuantumProvider,
			&attestation.QubitCount, &attestation.ErrorRate, &attestation.Status,
			&attestationData, &attestation.AttestedAt, &attestation.ExpiresAt,
		)
		if err != nil {
			continue
		}
		_ = json.Unmarshal(attestationData, &attestation.AttestationData)
		attestations = append(attestations, attestation)
	}

	c.JSON(http.StatusOK, gin.H{"attestations": attestations})
}

func (h *QuantumAttestationHandler) CreateAttestation(c *gin.Context) {
	tenantID := c.GetHeader("X-Tenant-ID")

	var attestation QuantumAttestation
	if err := c.ShouldBindJSON(&attestation); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	attestation.ID = uuid.New()
	attestation.TenantID, _ = uuid.Parse(tenantID)
	attestation.Status = "pending"

	attestationDataJSON, _ := json.Marshal(attestation.AttestationData)

	query := `
		INSERT INTO quantum_attestations 
		(id, tenant_id, asset_id, attestation_type, quantum_network, quantum_provider,
		 qubit_count, error_rate, status, attestation_data)
		VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10)
	`
	_, err := h.db.Exec(query,
		attestation.ID, attestation.TenantID, attestation.AssetID,
		attestation.AttestationType, attestation.QuantumNetwork, attestation.QuantumProvider,
		attestation.QubitCount, attestation.ErrorRate, attestation.Status, attestationDataJSON,
	)
	if err != nil {
		h.logger.Error("Failed to create attestation: ", err)
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to create attestation"})
		return
	}

	h.logger.Info("Created quantum attestation: ", attestation.ID)

	c.JSON(http.StatusCreated, attestation)
}

func (h *QuantumAttestationHandler) GetAttestation(c *gin.Context) {
	id := c.Param("id")
	attestationID, err := uuid.Parse(id)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid attestation ID"})
		return
	}

	var attestation QuantumAttestation
	var attestationData []byte
	query := `
		SELECT id, tenant_id, asset_id, attestation_type, quantum_network, quantum_provider,
		       qubit_count, error_rate, status, attestation_data, attested_at, expires_at
		FROM quantum_attestations WHERE id = $1
	`
	err = h.db.QueryRow(query, attestationID).Scan(
		&attestation.ID, &attestation.TenantID, &attestation.AssetID,
		&attestation.AttestationType, &attestation.QuantumNetwork, &attestation.QuantumProvider,
		&attestation.QubitCount, &attestation.ErrorRate, &attestation.Status,
		&attestationData, &attestation.AttestedAt, &attestation.ExpiresAt,
	)
	if err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": "Attestation not found"})
		return
	}

	_ = json.Unmarshal(attestationData, &attestation.AttestationData)

	c.JSON(http.StatusOK, attestation)
}

func (h *QuantumAttestationHandler) VerifyAttestation(c *gin.Context) {
	id := c.Param("id")
	attestationID, err := uuid.Parse(id)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid attestation ID"})
		return
	}

	h.logger.Info("Verifying quantum attestation: ", attestationID)

	query := `UPDATE quantum_attestations SET status = 'attested', attested_at = NOW() WHERE id = $1`
	_, _ = h.db.Exec(query, attestationID)

	c.JSON(http.StatusOK, gin.H{
		"id":          attestationID,
		"status":      "attested",
		"verified_at": time.Now(),
		"message":     "Attestation verified successfully",
	})
}

func (h *QuantumAttestationHandler) RefreshAttestation(c *gin.Context) {
	id := c.Param("id")
	attestationID, err := uuid.Parse(id)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid attestation ID"})
		return
	}

	h.logger.Info("Refreshing quantum attestation: ", attestationID)

	query := `UPDATE quantum_attestations SET status = 'pending', attested_at = NULL WHERE id = $1`
	_, _ = h.db.Exec(query, attestationID)

	c.JSON(http.StatusOK, gin.H{
		"id":      attestationID,
		"status":  "refreshed",
		"message": "Attestation refresh initiated",
	})
}

func (h *QuantumAttestationHandler) ListQuantumNetworks(c *gin.Context) {
	networks := []QuantumNetwork{
		{
			ID:        "ibmq_brisbane",
			Name:      "IBM Quantum Brisbane",
			Provider:  "IBM Quantum",
			Status:    "online",
			Qubits:    127,
			ErrorRate: 0.001,
			Fidelity:  0.99,
			Region:    "au-syd",
		},
		{
			ID:        "ibmq_sherbrooke",
			Name:      "IBM Quantum Sherbrooke",
			Provider:  "IBM Quantum",
			Status:    "online",
			Qubits:    127,
			ErrorRate: 0.001,
			Fidelity:  0.99,
			Region:    "ca-montreal",
		},
		{
			ID:        "ibm_washington",
			Name:      "IBM Quantum Washington",
			Provider:  "IBM Quantum",
			Status:    "online",
			Qubits:    127,
			ErrorRate: 0.001,
			Fidelity:  0.99,
			Region:    "us-east",
		},
	}

	c.JSON(http.StatusOK, gin.H{"networks": networks})
}

// demoQuantumProviders is the static fallback shown when the provider registry
// has no instantiated providers (e.g. registry initialisation skipped).
var demoQuantumProviders = []map[string]interface{}{
	{
		"name":      "IBM Quantum",
		"type":      "quantum_computing",
		"api_type":  "rest",
		"endpoints": []string{"https://api.quantum-computing.ibm.com"},
		"features":  []string{"Qiskit", "Quantum Circuits", "Hybrid Jobs"},
	},
	{
		"name":      "Q-CTRL",
		"type":      "quantum_control",
		"api_type":  "rest",
		"endpoints": []string{"https://api.q-ctrl.com"},
		"features":  []string{"Error Suppression", "Error Mitigation", "Pulse Level"},
	},
	{
		"name":      "Rigetti",
		"type":      "quantum_computing",
		"api_type":  "rest",
		"endpoints": []string{"https://api.rigetti.com"},
		"features":  []string{"Quil", "Quantum VM", "Hybrid Computing"},
	},
}

// ListQuantumProviders reflects the providers registered in the Quantum SDK
// provider registry, each with its live status and capabilities.
func (h *QuantumAttestationHandler) ListQuantumProviders(c *gin.Context) {
	providers := make([]gin.H, 0, len(h.providers.SortedProviderNames()))
	for _, name := range h.providers.SortedProviderNames() {
		p, ok := h.providers.Get(name)
		if !ok {
			continue
		}
		info := p.Info()
		providers = append(providers, gin.H{
			"name":          info.Name,
			"version":       info.Version,
			"vendor":        info.Vendor,
			"type":          info.Kind,
			"features":      info.Capabilities,
			"opt_in":        info.OptIn,
			"status":        p.Status(c.Request.Context()),
			"documentation": info.Documentation,
		})
	}

	if len(providers) == 0 {
		for _, demo := range demoQuantumProviders {
			providers = append(providers, gin.H(demo))
		}
	}

	c.JSON(http.StatusOK, gin.H{"providers": providers, "total": len(providers)})
}

// GetQuantumProviderStatus returns the live status of a single registered
// provider by name.
func (h *QuantumAttestationHandler) GetQuantumProviderStatus(c *gin.Context) {
	name := c.Param("name")
	p, ok := h.providers.Get(name)
	if !ok {
		c.JSON(http.StatusNotFound, gin.H{"error": "provider not found"})
		return
	}
	c.JSON(http.StatusOK, gin.H{
		"provider": p.Info().Name,
		"status":   p.Status(c.Request.Context()),
	})
}

func (h *QuantumAttestationHandler) GetQuantumReadiness(c *gin.Context) {
	tenantID := c.GetHeader("X-Tenant-ID")

	h.logger.Info("Calculating quantum readiness for tenant: ", tenantID)

	readiness := QuantumReadiness{
		OverallScore:      65,
		QuantumSafeAssets: 35,
		AtRiskAssets:      65,
		MigrationProgress: 35,
		Recommendations: []string{
			"Migrate RSA-2048 to CRYSTALS-Kyber",
			"Migrate ECDSA to CRYSTALS-Dilithium",
			"Implement hybrid classical-quantum key exchange",
			"Review cryptographic inventory for deprecated algorithms",
		},
		ByCategory: map[string]int{
			"symmetric_encryption":  80,
			"asymmetric_encryption": 45,
			"digital_signatures":    40,
			"key_exchange":          35,
			"hash_functions":        90,
		},
		ByAlgorithm: map[string]int{
			"RSA":                30,
			"ECDSA":              40,
			"ECDH":               35,
			"AES-256":            85,
			"SHA-256":            90,
			"CRYSTALS-Kyber":     10,
			"CRYSTALS-Dilithium": 5,
		},
	}

	c.JSON(http.StatusOK, readiness)
}

func (h *QuantumAttestationHandler) GetMigrationPlan(c *gin.Context) {
	tenantID := c.GetHeader("X-Tenant-ID")

	h.logger.Info("Generating migration plan for tenant: ", tenantID)

	migrationPlan := gin.H{
		"current_state": gin.H{
			"total_crypto_assets": 100,
			"quantum_safe":        35,
			"at_risk":             65,
		},
		"target_state": gin.H{
			"total_crypto_assets": 100,
			"quantum_safe":        100,
			"at_risk":             0,
		},
		"phases": []gin.H{
			{
				"phase":      1,
				"name":       "Assessment & Planning",
				"duration":   "2 weeks",
				"activities": []string{"Inventory crypto assets", "Risk assessment", "Prioritization"},
				"completion": "100%",
			},
			{
				"phase":      2,
				"name":       "Hybrid Migration",
				"duration":   "4 weeks",
				"activities": []string{"Implement hybrid RSA+Kyber", "Deploy ECDSA+Dilithium", "Update TLS config"},
				"completion": "60%",
			},
			{
				"phase":      3,
				"name":       "Full PQC Migration",
				"duration":   "6 weeks",
				"activities": []string{"Remove legacy algorithms", "Migrate key management", "Update applications"},
				"completion": "0%",
			},
		},
		"estimated_completion":    "12 weeks",
		"total_assets_to_migrate": 65,
	}

	c.JSON(http.StatusOK, migrationPlan)
}

func (h *QuantumAttestationHandler) ListPQCAlgorithms(c *gin.Context) {
	algorithms := []map[string]interface{}{
		{
			"name":            "CRYSTALS-Kyber",
			"type":            "KEM",
			"nist_round":      4,
			"security_level":  5,
			"key_size":        1568,
			"ciphertext_size": 1568,
			"status":          "standardized",
		},
		{
			"name":            "CRYSTALS-Dilithium",
			"type":            "Digital Signature",
			"nist_round":      4,
			"security_level":  5,
			"public_key_size": 2592,
			"signature_size":  4595,
			"status":          "standardized",
		},
		{
			"name":            "FALCON",
			"type":            "Digital Signature",
			"nist_round":      4,
			"security_level":  5,
			"public_key_size": 897,
			"signature_size":  666,
			"status":          "standardized",
		},
		{
			"name":            "SPHINCS+",
			"type":            "Digital Signature",
			"nist_round":      4,
			"security_level":  5,
			"public_key_size": 64,
			"signature_size":  49856,
			"status":          "standardized",
		},
		{
			"name":           "ML-KEM",
			"type":           "KEM",
			"nist_round":     4,
			"security_level": 5,
			"key_size":       1568,
			"status":         "standardized",
		},
		{
			"name":           "ML-DSA",
			"type":           "Digital Signature",
			"nist_round":     4,
			"security_level": 5,
			"status":         "standardized",
		},
	}

	c.JSON(http.StatusOK, gin.H{"algorithms": algorithms})
}

func (h *QuantumAttestationHandler) MigrateAlgorithm(c *gin.Context) {
	tenantID := c.GetHeader("X-Tenant-ID")

	var req struct {
		AssetID         string `json:"asset_id"`
		SourceAlgorithm string `json:"source_algorithm"`
		TargetAlgorithm string `json:"target_algorithm"`
	}
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	h.logger.Info("Migrating algorithm for tenant: ", tenantID, " from: ", req.SourceAlgorithm, " to: ", req.TargetAlgorithm)

	c.JSON(http.StatusOK, gin.H{
		"status":             "migration_started",
		"source_algorithm":   req.SourceAlgorithm,
		"target_algorithm":   req.TargetAlgorithm,
		"estimated_duration": "5 minutes",
		"message":            "Algorithm migration initiated",
	})
}

type IBMQuantumClient struct {
	APIKey  string
	BaseURL string
	Network string
	Timeout int
}

func NewIBMQuantumClientForAttestation(apiKey, baseURL, network string) *IBMQuantumClient {
	return &IBMQuantumClient{
		APIKey:  apiKey,
		BaseURL: baseURL,
		Network: network,
		Timeout: 30,
	}
}

func (c *IBMQuantumClient) GetNetworkInfo(ctx context.Context) (map[string]interface{}, error) {
	return map[string]interface{}{
		"name":     c.Network,
		"nodes":    5,
		"qubits":   127,
		"fidelity": 0.994,
		"status":   "online",
	}, nil
}

func (c *IBMQuantumClient) SubmitJob(ctx context.Context, job map[string]interface{}) (string, error) {
	return uuid.New().String(), nil
}

func (c *IBMQuantumClient) GetJobResult(ctx context.Context, jobID string) (map[string]interface{}, error) {
	return map[string]interface{}{
		"job_id":      jobID,
		"status":      "completed",
		"result":      " attestation verified",
		"qubits_used": 127,
		"error_rate":  0.001,
	}, nil
}

func (c *IBMQuantumClient) VerifyAttestation(ctx context.Context, data map[string]interface{}) (bool, error) {
	return true, nil
}